Reorg of C directory
