+++ implementations/c/build.gradle
@@ -6,39 +6,13 @@ plugins {
 
 task build {
   onlyIf { host.debianBuilder.enabled }
-  doLast {
-    builderExec 'debian', {
-      script '''
-        mkdir -p _build/x86_64-unknown-linux-gnu
-        cd _build/x86_64-unknown-linux-gnu
-        cmake ../.. -DCMAKE_INSTALL_PREFIX=./_install
-        make
-        make install
-      '''
-    }
-  }
+  dependsOn gradle.includedBuild('testatecc508a').task(':build')
 }
 
 task test {
-  dependsOn gradle.includedBuild('test').task(':test')
+  dependsOn gradle.includedBuild('testatecc508a').task(':test')
 }
 
 task clean {
-  doLast {
-    delete '_build'
-  }
+  dependsOn gradle.includedBuild('testatecc508a').task(':clean')
 }
-
-// mkdir -p _build/arm-unknown-linux-gnueabi
-// cd _build/arm-unknown-linux-gnueabi
-// cmake ../.. -DCMAKE_TOOLCHAIN_FILE=../../toolchain/arm-unknown-linux-gnueabi.cmake -DCMAKE_INSTALL_PREFIX=./_install
-// make
-// make install
-// cd -
-//
-// mkdir -p _build/x86_64-unknown-linux-gnu
-// cd _build/x86_64-unknown-linux-gnu
-// cmake ../.. -DCMAKE_TOOLCHAIN_FILE=../../toolchain/x86_64-unknown-linux-gnu.cmake -DCMAKE_INSTALL_PREFIX=./_install
-// make
-// make install
-// cd -
