+++ implementations/c/config/vault/ockam_vault_hw_cfg.h
