+++ implementations/c/include/ockam_def.h
