+++ implementations/c/include/ockam_err.h
