+++ implementations/c/include/vault/ockam_vault_def.h
