+++ implementations/c/include/vault/ockam_vault_hw_microchip.h
