--- implementations/c/ockam/common/inc/ockam_mem.h
@@ -1,98 +1,0 @@
-/**
- ********************************************************************************************************
- * @file        ockam_mem.h
- * @author      Mark Mulrooney <mark@ockam.io>
- * @copyright   Copyright (c) 2019, Ockam Inc.
- * @brief   
- ********************************************************************************************************
- */
-
-#ifndef OCKAM_MEM_H_
-#define OCKAM_MEM_H_
-
-
-/*
- ********************************************************************************************************
- * @defgroup    OCKAM_MEM OCKAM_MEM_API
- * @ingroup     OCKAM
- * @brief       OCKAM_MEM_API
- *
- * @addtogroup  OCKAM_MEM
- * @{
- ********************************************************************************************************
- */
-
-
-/*
- ********************************************************************************************************
- *                                             INCLUDE FILES                                            *
- ********************************************************************************************************
- */
-
-#include <common/inc/ockam_def.h>
-
-
-/*
- ********************************************************************************************************
- *                                                DEFINES                                               *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                               CONSTANTS                                              *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                               DATA TYPES                                             *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                          FUNCTION PROTOTYPES                                         *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                            GLOBAL VARIABLES                                          *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                           GLOBAL FUNCTIONS                                           *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                            LOCAL FUNCTIONS                                           *
- ********************************************************************************************************
- */
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-OCKAM_ERR ockam_mem_init(void* p_buf);
-
-OCKAM_ERR ockam_mem_alloc(void* p_buf, uint32_t size);
-
-OCKAM_ERR ockam_mem_free(void* p_buf);
-
-
-#ifdef __cplusplus
-}
-#endif
-
-/*
- ********************************************************************************************************
- * @}
- ********************************************************************************************************
- */
-
-#endif
