--- implementations/c/ockam/common/src/kal/ockam_kal_linux.c
@@ -1,146 +1,0 @@
-/**
- ********************************************************************************************************
- * @file        ockam_kal_linux.c
- * @author      Mark Mulrooney <mark@ockam.io>
- * @copyright   Copyright (c) 2019, Ockam Inc.
- * @brief   
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                             INCLUDE FILES                                            *
- ********************************************************************************************************
- */
-
-#include <common/inc/ockam_err.h>
-#include <common/inc/ockam_kal.h>
-
-
-/*
- ********************************************************************************************************
- *                                                DEFINES                                               *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                               CONSTANTS                                              *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                               DATA TYPES                                             *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                          FUNCTION PROTOTYPES                                         *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                            GLOBAL VARIABLES                                          *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                           GLOBAL FUNCTIONS                                           *
- ********************************************************************************************************
- */
-
-/*
- ********************************************************************************************************
- *                                            LOCAL FUNCTIONS                                           *
- ********************************************************************************************************
- */
-
-/**
- ********************************************************************************************************
- *                                          ockam_kal_mutex_init()
- *
- * @brief   Initialize a mutex
- *
- * @param   p_mutex The Mutex object to initialize
- * 
- * @return  OCKAM_ERR_NONE on success.
- *
- ********************************************************************************************************
- */
-
-OCKAM_ERR ockam_kal_mutex_init(OCKAM_KAL_MUTEX *p_mutex)
-{
-    return OCKAM_ERR_NONE;
-}
-
-
-/**
- ********************************************************************************************************
- *                                          ockam_kal_mutex_free()
- *
- * @brief   Free a mutex
- *
- * @param   p_mutex The Mutex object to free 
- * 
- * @return  OCKAM_ERR_NONE on success.
- *
- ********************************************************************************************************
- */
-
-OCKAM_ERR ockam_kal_mutex_free(OCKAM_KAL_MUTEX *p_mutex)
-{
-    return OCKAM_ERR_NONE;
-}
-
-
-/**
- ********************************************************************************************************
- *                                          ockam_kal_mutex_lock()
- *
- * @brief   Lock the specified mutex
- *
- * @param   p_mutex     The mutex to lock.
- *
- * @param   opt         Options to pass into the mutex lock .
- *
- * @param   timeout_ms  The maximum amount of time to wait for the mutex.
- * 
- * @return  OCKAM_ERR_NONE when mutex is acquired.
- *
- ********************************************************************************************************
- */
-
-OCKAM_ERR ockam_kal_mutex_lock(OCKAM_KAL_MUTEX *p_mutex,
-                               OCKAM_KAL_OPT opt,
-                               uint32_t timeout_ms)
-{
-    return OCKAM_ERR_NONE;
-}
-
-
-
-/**
- ********************************************************************************************************
- *                                          ockam_kal_mutex_unlock()
- *
- * @brief   Free the specified mutex
- *
- * @param   p_mutex The mutex object to release.
- *
- * @param   opt     Options to pass to the mutex unlock.
- * 
- * @return  OCKAM_ERR_NONE after the mutex is freed.
- *
- ********************************************************************************************************
- */
-
-OCKAM_ERR ockam_kal_mutex_unlock (OCKAM_KAL_MUTEX *p_mutex,
-                                  OCKAM_KAL_OPT opt)
-{
-    return OCKAM_ERR_NONE;
-}
-
