+++ implementations/c/settings.gradle
@@ -5,5 +5,5 @@ boolean inComposite = gradle.parent != null
 if (!inComposite) {
   includeBuild '../../tools/gradle/plugins/host'
   includeBuild '../../tools/gradle/plugins/builder'
-  includeBuild 'test'
+  includeBuild 'test/vault/atecc508a'
 }
