+++ implementations/c/source/log/ockam_log_linux.c
@@ -1,0 +1,103 @@
+/**
+ ********************************************************************************************************
+ * @file        ockam_log_linux.c
+ * @author      Mark Mulrooney <mark@ockam.io>
+ * @copyright   Copyright (c) 2019, Ockam Inc.
+ * @brief   
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                             INCLUDE FILES                                            *
+ ********************************************************************************************************
+ */
+
+#include <ockam_def.h>
+#include <ockam_err.h>
+
+#include <log/ockam_log.h>
+
+
+/*
+ ********************************************************************************************************
+ *                                                DEFINES                                               *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                               CONSTANTS                                              *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                               DATA TYPES                                             *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                          FUNCTION PROTOTYPES                                         *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                            GLOBAL VARIABLES                                          *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                           GLOBAL FUNCTIONS                                           *
+ ********************************************************************************************************
+ */
+
+/*
+ ********************************************************************************************************
+ *                                            LOCAL FUNCTIONS                                           *
+ ********************************************************************************************************
+ */
+
+
+/**
+ ********************************************************************************************************
+ *                                            ockam_log_init()
+ *
+ * @brief   Initialize the Ockam generic log functions.
+ *
+ * @return  OCKAM_ERR_NONE on success.
+ * 
+ ********************************************************************************************************
+ */
+
+OCKAM_ERR ockam_log_init(void)
+{
+    return OCKAM_ERR_NONE;
+}
+
+
+/**
+ ********************************************************************************************************
+ *                                             ockam_log()
+ *
+ * @brief   Ockam's generic log function
+ *
+ * @param   p_str[in]       String buffer to write to the log
+ *
+ * @param   str_size[in]    Size of the string to write to the log
+ * 
+ * @return  OCKAM_ERR_NONE on success.
+ * 
+ ********************************************************************************************************
+ */
+
+OCKAM_ERR ockam_log(void* p_str, uint32_t str_size)
+{
+    return OCKAM_ERR_NONE;
+}
+
+
+
