--- implementations/c/test/settings.gradle
@@ -1,7 +1,0 @@
-rootProject.name = 'test'
-
-boolean inComposite = gradle.parent != null
-if (!inComposite) {
-  includeBuild '../../../tools/gradle/plugins/host'
-  includeBuild '../../../tools/gradle/plugins/builder'
-}
