+++ implementations/c/test/vault/atecc508a/build.gradle
@@ -1,0 +1,50 @@
+
+plugins {
+  id 'network.ockam.gradle.host' version '1.0.0'
+  id 'network.ockam.gradle.builders' version '1.0.0'
+}
+
+task x {
+    doLast {
+        builderExec 'debian', {
+            script '''
+                mkdir -p _build/
+                cd _build/
+                cmake .. 
+                make
+                '''
+        }
+    }
+}
+
+task build {
+  onlyIf { host.debianBuilder.enabled }
+  doLast {
+    builderExec 'debian', {
+      script '''
+        mkdir -p _build/
+        cd _build/
+        cmake .. 
+        make
+      '''
+    }
+  }
+}
+
+task test {
+  dependsOn build
+  onlyIf { host.debianBuilder.enabled }
+  //doLast {
+  //  builderExec 'debian', {
+  //    script '''
+  //      ./_build/x86_64-unknown-linux-gnu/_install/bin/test_ockam
+  //    '''
+  //  }
+  //}
+}
+
+task clean {
+  doLast {
+    delete '_build'
+  }
+}
