+++ implementations/c/test/vault/atecc508a/config/ockam_vault_cfg.h
@@ -1,0 +1,41 @@
+/**
+ ********************************************************************************************************
+ * @file        ockam_vault_cfg.h
+ * @author      Mark Mulrooney <mark@ockam.io>
+ * @copyright   Copyright (c) 2019, Ockam Inc.
+ * @brief   
+ ********************************************************************************************************
+ */
+
+#ifndef OCKAM_VAULT_CFG_H_
+#define OCKAM_VAULT_CFG_H_
+
+
+/*
+ ********************************************************************************************************
+ *                                               INCLUDES                                               *
+ ********************************************************************************************************
+ */
+
+#include <vault/ockam_vault_def.h>
+
+
+/*
+ ********************************************************************************************************
+ *                                         Function Configuration                                       *
+ ********************************************************************************************************
+ */
+
+
+#define OCKAM_VAULT_CFG_INIT               OCKAM_VAULT_HW_MICROCHIP_ATECC508A | OCKAM_VAULT_SW_MBEDTLS
+
+#define OCKAM_VAULT_CFG_RAND               OCKAM_VAULT_HW_MICROCHIP_ATECC508A
+
+#define OCKAM_VAULT_CFG_KEY_ECDH           OCKAM_VAULT_HW_MICROCHIP_ATECC508A
+
+#define OCKAM_VAULT_CFG_HKDF               OCKAM_VAULT_SW_MBEDTLS
+
+#define OCKAM_VAULT_CFG_AES_GCM            OCKAM_VAULT_SW_MBEDTLS
+
+
+#endif
