+++ implementations/c/test/vault/atecc508a/settings.gradle
@@ -1,0 +1,7 @@
+rootProject.name = 'testatecc508a'
+
+boolean inComposite = gradle.parent != null
+if (!inComposite) {
+  includeBuild '../../../../../tools/gradle/plugins/host'
+  includeBuild '../../../../../tools/gradle/plugins/builder'
+}
