--- implementations/c/test/vault/build.gradle
@@ -1,38 +1,0 @@
-
-plugins {
-  id 'network.ockam.gradle.host' version '1.0.0'
-  id 'network.ockam.gradle.builders' version '1.0.0'
-}
-
-task build {
-  onlyIf { host.debianBuilder.enabled }
-  doLast {
-    builderExec 'debian', {
-      script '''
-        mkdir -p _build/x86_64-unknown-linux-gnu
-        cd _build/x86_64-unknown-linux-gnu
-        cmake ../.. -DCMAKE_INSTALL_PREFIX=./_install
-        make
-        make install
-      '''
-    }
-  }
-}
-
-task test {
-  dependsOn build
-  onlyIf { host.debianBuilder.enabled }
-  doLast {
-    builderExec 'debian', {
-      script '''
-        ./_build/x86_64-unknown-linux-gnu/_install/bin/test_ockam
-      '''
-    }
-  }
-}
-
-task clean {
-  doLast {
-    delete '_build'
-  }
-}
